Added a lot of things
NEW: Added Fallback sounds to prevent crashes when using it
NEW: Added Fonts to prevent default fonts to be used
Also including with this is a fix for foreign fonts, that brings MOST letters back to OITG. some are still broken.
NEW: Made some choices for TitleScreen that are available in Graphics, so the default ones don't interfere.