Revision 363
- Fixed Assist Tick soundeffect
- Removed an unnecessary copy of a file that was there
- Added quite a lot of features that are included in the lua file.
**Enable Characters in Player Options**
Change this if you want to see the character list on Player Options. Some people might want to use this if they have the character screen disabled, yet they want to use them, but there's also people who just don't want to use them at all. So this options is for them.

**Use ITG Noteskin System**
Change this if you want to use the limited Noteskin system that the default OITG theme uses. If you want to add more to that list, you have to edit the 'english.ini' file from your DEFAULT theme in order to make it work.

**Enable Score Display Listing**
Change this if you want to enable or disable the Score Display methods found in Player Options. Here's a little thing: I don't know someone who uses this thing. So, here's an option to enable it or disable it if you wish to.

**Restore Internet Listing**
Although an uncesserary option, you can "re-enable" it by changing this option. (I DO NOT guarantee that Net Functionality works correctly.)

**Advanced Graphic Options**
Enabiling this option will let you access a more advanced and flexed out Graphic system that can let you change aspect ratio, resolution, video mode, coloring, etc.::Some options from other menus will disable because they'll move to this menu when you enable it.

**Height Position For Receptors**
Change the height of the receptors to a preset. Some themes have different heights, and even some stepfiles break when they're not on a appropriate height, so some options are available.

*GLOBAL ON THEME*: Allows the framework to change the height depending on the theme's original receptor position.

*Cyberia-Style / A.o.i*: Sets the position to the ones found in Cyberia Style and A.o.I.

*DISABLE*: Resets the position of the receptors, and goes to the original position found in OpenITG.
The value found in GLOBAL ON THEME updates when re-entering the screen.

- Made a warning informing about Course Edit not exiting on OITG.
- Added credit to Mark Cannon ("Vyhd") for his OpenITG simple graphics configuration - version 0.2, OpenITG resolution switcher - version 1.0 and Simple PrefsRows API for 3.95/OpenITG, version 1.2.